[IMP] apply default theme after installation +
reload themes on installing new modules with customize_show

+++ website_multi_theme/models/website.py
@@ -86,7 +86,9 @@ class Website(models.Model):
             return result
         # Copy patterns only for current website
         key = xmlid if override_key else pattern.key
-        result = pattern.copy({
+        result = pattern.with_context(
+            duplicate_view_for_website=True
+        ).copy({
             "active": pattern.was_active,
             "arch_fs": False,
             "key": key,
@@ -124,15 +126,21 @@ class Website(models.Model):
         assets_pattern = self.env.ref("website_multi_theme.assets_pattern")
         layout_pattern = self.env.ref("website_multi_theme.layout_pattern")
         for website in self:
-            # Websites without multi theme need to clean their previous views
             if not website.multi_theme_id:
-                _logger.info(
-                    "Deleting multi website theme views for %s: %s",
-                    website.display_name,
-                    website.multi_theme_view_ids,
+                default_theme = self.env.ref(
+                    'website_multi_theme.theme_default'
                 )
-                website.multi_theme_view_ids.unlink()
-                continue
+                if not default_theme:
+                    _logger.info(
+                        "Deleting multi website theme views for %s: %s",
+                        website.display_name,
+                        website.multi_theme_view_ids,
+                    )
+                    website.multi_theme_view_ids.unlink()
+                    continue
+                else:
+                    website.multi_theme_id = default_theme
+
             # Duplicate multi theme patterns for this website
             custom_assets = website._duplicate_view_for_website(
                 assets_pattern,
